# docs/swagger.yaml
basePath: /api/v1
definitions:
  api.ErrorResp:
    properties:
      message:
        type: string
    type: object
...
  dto.AdminCategoryDetail:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/dto.ProductListItem'
        type: array
      published:
        type: boolean
      slug:
        type: string
      updatedAt:
        type: string
    type: object
  dto.ApiError:
    properties:
      code:
        type: string
      details:
        type: string
      stack:
        description: Hide in production
    type: object
...
  dto.ApiResponse-array_dto_AdminCategoryDetail:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.AdminCategoryDetail'
        type: array
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-array_dto_CategoryDetail:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.CategoryDetail'
        type: array
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.ApiResponse-dto_CategoryDetail:
    properties:
      data:
        $ref: '#/definitions/dto.CategoryDetail'
      error:
        $ref: '#/definitions/dto.ApiError'
      message:
        type: string
      meta:
        $ref: '#/definitions/dto.MetaInfo'
      pagination:
        $ref: '#/definitions/dto.Pagination'
    type: object
...
  dto.CategoryDetail:
    properties:
      createdAt:
        type: string
      description:
        type: string
      id:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      products:
        items:
          $ref: '#/definitions/dto.ProductSummary'
        type: array
      published:
        type: boolean
      slug:
        type: string
    type: object
...
  dto.ErrorResp:
    properties:
      error:
        $ref: '#/definitions/dto.ApiError'
    type: object
...
  dto.MetaInfo:
    properties:
      method:
        type: string
      path:
        type: string
      requestId:
        type: string
      timestamp:
        type: string
    type: object
...
  dto.Pagination:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      page:
        type: integer
      pageSize:
        type: integer
      total:
        type: integer
      totalPages:
        type: integer
    type: object
...
  dto.ProductListItem:
    properties:
      avgRating:
        type: number
      basePrice:
        type: number
      createdAt:
        type: string
      description:
        type: string
      discountPercentage:
        type: integer
      id:
        type: string
      imageId:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      reviewCount:
        type: integer
      sku:
        type: string
      slug:
        type: string
      updatedAt:
        type: string
    type: object
...
  dto.ProductSummary:
    properties:
      avgRating:
        type: number
      createdAt:
        type: string
      id:
        type: string
      imageId:
        type: string
      imageUrl:
        type: string
      name:
        type: string
      price:
        type: number
      reviewCount:
        type: integer
      slug:
        type: string
      updatedAt:
        type: string
      variantCount:
        type: integer
    type: object
...
  models.CreateCategoryModel:
    properties:
      description:
        maxLength: 1000
        type: string
      displayOrder:
        type: integer
      image:
        $ref: '#/definitions/multipart.FileHeader'
      name:
        maxLength: 255
        minLength: 3
        type: string
      slug:
        maxLength: 255
        minLength: 3
        type: string
    required:
    - name
    - slug
    type: object
...
  models.UpdateCategoryModel:
    properties:
      description:
        maxLength: 1000
        type: string
      displayOrder:
        type: integer
      image:
        $ref: '#/definitions/multipart.FileHeader'
      name:
        maxLength: 255
        minLength: 3
        type: string
      published:
        type: boolean
      slug:
        maxLength: 255
        minLength: 3
        type: string
    type: object
...
  multipart.FileHeader:
    properties:
      filename:
        type: string
      header:
        $ref: '#/definitions/textproto.MIMEHeader'
      size:
        format: int64
        type: integer
    type: object
...
  textproto.MIMEHeader:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
host: localhost:4000
info:
  contact: {}
  description: This is a sample server for a simple e-commerce API.
  title: E-Commerce API
paths:
...
  /admin/categories:
    get:
      consumes:
      - application/json
      description: Get a list of Categories
      operationId: get-admin-Categories
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-array_dto_CategoryDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Get a list of Categories
      tags:
      - Categories
    post:
      consumes:
      - application/json
      description: Create a new Category
      operationId: create-Category
      parameters:
      - description: Category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateCategoryModel'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_CategoryDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Create a new Category
      tags:
      - Categories
  /admin/categories/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a Category
      operationId: delete-Category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Delete a Category
      tags:
      - Admin
    get:
      consumes:
      - application/json
      description: Get a Category by ID
      operationId: get-Category-by-id
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_CategoryDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Get a Category by ID
      tags:
      - Categories
    put:
      consumes:
      - application/json
      description: Update a Category
      operationId: update-Category
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: integer
      - description: Category request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateCategoryModel'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_CategoryDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResp'
      summary: Update a Category
      tags:
      - Admin
...
  /categories:
    get:
      consumes:
      - application/json
      description: Get a list of Categories
      operationId: get-Categories
      parameters:
      - description: Page number
        in: query
        name: page
        type: integer
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-array_dto_AdminCategoryDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Get a list of Categories
      tags:
      - Categories
  /categories/slug/{slug}:
    get:
      consumes:
      - application/json
      description: Get a list of Products by Category Slug
      operationId: get-Products-by-Category-Slug
      parameters:
      - description: Category Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Page size
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ApiResponse-dto_CategoryDetail'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/api.ErrorResp'
      summary: Get a list of Products by Category Slug
      tags:
      - Categories
...
swagger: "2.0"